requirements.txt
flask>=2.3.0
flask-restx>=0.5.1
numpy>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
requests>=2.31.0
werkzeug>=2.3.0
streamlit>=1.25.0
python-dotenv>=1.0.0
azure-ai-projects>=1.0.0
azure-identity>=1.14.0
pytest>=7.4.0
